set server routes
