+++ wireguard/wireguard_linux.go
@@ -7,9 +7,7 @@ import (
 	"os/exec"
 	"strings"
 
-	"github.com/gravitl/netclient/config"
 	"github.com/gravitl/netclient/ncutils"
-	"github.com/gravitl/netclient/stun"
 	"github.com/gravitl/netmaker/logger"
 	"github.com/vishvananda/netlink"
 	"golang.org/x/exp/slog"
@@ -309,57 +307,6 @@ func DeleteOldInterface(iface string) {
 	if _, err := ncutils.RunCmd(ip+" link del "+iface, true); err != nil {
 		logger.Log(0, "error removing interface", iface, err.Error())
 	}
-}
-
-func GetServerAddressesDefaultGw(server *config.Server) (addrs []net.IPNet) {
-	if server == nil {
-		return
-	}
-	ips, _ := net.LookupIP(server.Name) // handle server base domain
-	for _, ip := range ips {
-		ipnet := GetIPNetfromIp(ip)
-		if ipnet != nil {
-			addrs = append(addrs, *ipnet)
-		}
-	}
-
-	ips, _ = net.LookupIP(server.API) // handle server api
-	for _, ip := range ips {
-		ipnet := GetIPNetfromIp(ip)
-		if ipnet != nil {
-			addrs = append(addrs, *ipnet)
-		}
-	}
-
-	broker := server.Broker
-	brokerParts := strings.Split(broker, "//")
-	if len(brokerParts) > 1 {
-		broker = brokerParts[1]
-	}
-
-	ips, _ = net.LookupIP(broker) // handle server broker
-	for _, ip := range ips {
-		ipnet := GetIPNetfromIp(ip)
-		if ipnet != nil {
-			addrs = append(addrs, *ipnet)
-		}
-	}
-
-	stunList := stun.StunServers
-	for i := range stunList {
-		stunServer := stunList[i]
-		ips, err := net.LookupIP(stunServer.Domain) // handle server broker
-		if err != nil {
-			continue
-		}
-		for _, ip := range ips {
-			ipnet := GetIPNetfromIp(ip)
-			if ipnet != nil {
-				addrs = append(addrs, *ipnet)
-			}
-		}
-	}
-	return
 }
 
 func SetNmServerRoutes(addrs []net.IPNet) error {
